fixed donation add donation issue

+++ routes/web.php
@@ -408,11 +408,9 @@ Route::middleware(['admin.auth', '2fa'])->group(function () use ($controller_pat
             Route::get('/categories', [CategoryController::class, 'index'])->name('organizations.categories.index');
         });
     // Donation
-    Route::resource('/donations', DonationController::class);
-    Route::post('/add_donation', [DonationController::class, 'add_donation'])->name('create.donation');
-    Route::put('/edit_donation/{id}', [DonationController::class, 'edit_donation'])->name('edit.donation');
-    Route::delete('/destroy_donation/{id}', [DonationController::class, 'destroy_donation'])->name('destroy.donation');
-
+    Route::post('/add_donation', [App\Http\Controllers\Admin\Donation\DonationController::class, 'add_donation'])->name('create.donation');
+    Route::put('/edit_donation/{id}', [App\Http\Controllers\Admin\Donation\DonationController::class, 'edit_donation'])->name('edit.donation');
+    Route::delete('/destroy_donation/{id}', [App\Http\Controllers\Admin\Donation\DonationController::class, 'destroy_donation'])->name('destroy.donation');
     //Organization
     Route::post('/add_organization', [App\Http\Controllers\Admin\Donation\OrganizationController::class, 'add_organization'])->name('add.organization');
     Route::delete('/delete_organization/{id}', [App\Http\Controllers\Admin\Donation\OrganizationController::class, 'organization_destroy'])->name('organization.destroy');
